This commit contains changes which should fix ipyvolume for JupyterLab 2.1.0 (#323)
In utils.ts, the version string was set to a valid version, and taken from the package.json. I believe this is the true fix.

labplugin.ts was modified to use the semver from utils.ts instead of index.ts (reason given below). I also added the type signatures to try and catch any typing errors, but I realise that you probably omitted these deliberately (to reduce the lib size when exporting?).

index.ts was modified to remove any non-view/model/exporter exports, so it may be directly used as the `exports` argument to `registerWidget`. I'm not sure if this solves anything - but I could forsee a case where `registerWidget` expects every member of `exports` to be a view/model/exporter class.

I added a version 2 `||` constraint to `@jupyter-widgets/controls`, because the latest version is 2. Not sure if this is correct.